requirements.txt
asgiref==3.4.1
Django==3.2.25
django-filter==21.1
gunicorn==21.2.0
importlib-metadata==4.8.3
packaging==21.3
pyparsing==3.1.4
pytz==2025.2
sqlparse==0.4.4
typing-extensions==4.1.1
zipp==3.6.0
psycopg2-binary>=2.9
openpyxl>=3.1
djangorestframework>=3.15
